[AutofillAi - M4] Refactor the EntityMetadata::mUseCount to long.
This is done to make the C++ and Java EntityInstance classes as
compatible as possible.

Bug: 509798874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7827685
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1629328}

+++ chrome/browser/autofill/android/entity_instance_android.cc
@@ -70,7 +70,7 @@ EntityInstanceAndroid::EntityInstanceAndroid(
       record_type(entity_instance.record_type()),
       metadata(entity_instance.guid().value(),
                entity_instance.date_modified(),
-               static_cast<int>(entity_instance.use_count()),
+               entity_instance.use_count(),
                entity_instance.use_date()),
       requires_reauth_to_see(requires_reauth_to_see),
       is_masked_server_entity(entity_instance.IsMaskedEntity() &&
