+++ chrome/browser/autofill/android/entity_metadata_android.h
@@ -5,6 +5,8 @@
 #ifndef CHROME_BROWSER_AUTOFILL_ANDROID_ENTITY_METADATA_ANDROID_H_
 #define CHROME_BROWSER_AUTOFILL_ANDROID_ENTITY_METADATA_ANDROID_H_
 
+#include <cstdint>
+
 #include "base/android/jni_android.h"
 #include "base/android/scoped_java_ref.h"
 #include "base/time/time.h"
@@ -25,7 +27,7 @@ struct EntityMetadataAndroid {
 
   EntityMetadataAndroid(std::string guid,
                         base::Time date_modified,
-                        int use_count,
+                        int64_t use_count,
                         base::Time use_date);
   EntityMetadataAndroid(const EntityMetadataAndroid&);
   EntityMetadataAndroid& operator=(const EntityMetadataAndroid&);
@@ -35,7 +37,7 @@ struct EntityMetadataAndroid {
 
   std::string guid;
   base::Time date_modified;
-  int use_count;
+  int64_t use_count;
   base::Time use_date;
 };
